pyproject.toml
[project]
name = "openai-sdk-app-builder-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai (>=1.76.0,<2.0.0)",
    "openai-agents (>=0.0.13,<0.0.14)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)"
]

[tool.poetry]
packages = [{include = "openai_sdk_app_builder_agent", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
